refactor: rename testutils package to test
