+++ plugins/extractors/gcs/gcs_test.go
@@ -8,13 +8,13 @@ import (
 
 	"github.com/odpf/meteor/plugins"
 	"github.com/odpf/meteor/plugins/extractors/gcs"
-	"github.com/odpf/meteor/testutils"
+	"github.com/odpf/meteor/test"
 	"github.com/stretchr/testify/assert"
 )
 
 func TestExtract(t *testing.T) {
 	t.Run("should return error if no project_id in config", func(t *testing.T) {
-		err := gcs.New(testutils.Logger).Extract(context.TODO(), map[string]interface{}{
+		err := gcs.New(test.Logger).Extract(context.TODO(), map[string]interface{}{
 			"wrong-config": "sample-project",
 		}, make(chan interface{}))
