+++ plugins/extractors/mssql/mssql_test.go
@@ -17,7 +17,7 @@ import (
 	"github.com/odpf/meteor/proto/odpf/assets"
 	"github.com/odpf/meteor/proto/odpf/assets/common"
 	"github.com/odpf/meteor/proto/odpf/assets/facets"
-	"github.com/odpf/meteor/testutils"
+	"github.com/odpf/meteor/test"
 	"github.com/ory/dockertest/v3"
 	"github.com/ory/dockertest/v3/docker"
 	"github.com/stretchr/testify/assert"
@@ -57,7 +57,7 @@ func TestMain(m *testing.M) {
 		}
 		return db.Ping()
 	}
-	err, purgeFn := testutils.CreateContainer(opts, retryFn)
+	err, purgeFn := test.CreateContainer(opts, retryFn)
 	if err != nil {
 		log.Fatal(err)
 	}
@@ -150,7 +150,7 @@ func execute(db *sql.DB, queries []string) (err error) {
 }
 
 func newExtractor() *mssql.Extractor {
-	return mssql.New(testutils.Logger)
+	return mssql.New(test.Logger)
 }
 
 func getExpected() (expected []assets.Table) {
