+++ test/logger.go
@@ -1,4 +1,4 @@
-package testutils
+package test
 
 import (
 	"io/ioutil"
